# api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: Odal Node API
  version: 0.12.0
  description: |
    **Sovereign Digital Product Passport Infrastructure**

    The Odal API is split across two deployable services:

    - **Node** (`http://localhost:8001`) — the MVP service that bundles vault,
      identity, and integrator on a single port with sub-path routing:
      `/vault/*`, `/identity/*`, `/integrator/*`.
    - **Resolver** (`http://localhost:8003`) — standalone public resolver for
      QR code scans and supply-chain integrations. Unauthenticated.

    All authenticated vault endpoints are versioned under `/vault/api/v1/`.
    Path parameters use OpenAPI 3.1 `{id}` brace syntax (aligned with
    Axum 0.8 routing).

    Authentication uses `Authorization: Bearer odal_sk_...` API keys (SHA-256
    hashed, prefix-indexed) or HTTP Basic auth for local development.

    Regulation references:
    - EU Ecodesign for Sustainable Products Regulation (ESPR) 2024/1781
    - EU Battery Regulation 2023/1542
    - W3C DID Core Spec 1.0 / W3C VC Data Model 2.0
  contact:
    name: Odal Support
    email: contact@odal-node.io
    url: https://docs.odal-node.io
  license:
    name: BSL-1.1
    url: https://mariadb.com/bsl11/
  x-legal-entity: Odal Node
servers:
  - url: http://localhost:8001
    description: Local development — odal-node (vault + identity public + integrator)
  - url: http://localhost:8002
...
  - url: http://localhost:8003
    description: Local development — odal-resolver (public)
...
paths:
...
  /dpp/{dppId}:
    get:
      operationId: resolveDpp
      summary: Resolve a public DPP
      description: |
        Resolve a published Digital Product Passport by ID. This endpoint
        is the target of QR code scans. No authentication required.

        Every representation is built from the **signed** public payload, not
        the live database row, so the body and the proof it carries agree by
        construction.

        Content negotiation via `Accept` header:
        - `application/json` / `application/ld+json` (default): JSON-LD
          passport data
        - `text/html`: the consumer-facing HTML passport page with operator
          branding (logo, colours)
        - `application/aas+json`: an IDTA Asset Administration Shell
          Environment (see below)

        An absent, empty, `*/*`, `application/*`, `application/json` or
        `application/ld+json` header all reach the JSON-LD default. Only a
        header naming something this route cannot produce gets `406`.

        Responses carry `Vary: Accept`.
      tags:
        - Public Resolver
      parameters:
        - name: dppId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DppId'
      responses:
        '200':
          description: DPP resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportResponse'
            text/html:
              schema:
                type: string
                description: Consumer-facing HTML passport page.
            application/aas+json:
              schema:
                type: object
                description: |
                  An IDTA Asset Administration Shell `Environment` — shells
                  and submodels in one self-contained document.

                  `conceptDescriptions` is **absent**, not empty. This node
                  coins no concept descriptions, and the metamodel constrains
                  that member to `minItems: 1`, so an empty array would make
                  the whole document invalid.

                  **Public tier only.** The passport is filtered through the
                  disclosure seam before any AAS mapper sees it, so this door
                  never carries a field the JSON-LD door would withhold.
                  Restricted and conformity-tier data require a credentialed
                  channel and a different projection.

                  **Schema-valid, not conformance-certified.** Every
                  Environment is validated in `dpp-core`'s CI against IDTA's
                  published AAS JSON Schemas for metamodel **3.0, 3.1 and
                  3.2**, and must satisfy all three — no single revision is
                  the strictest, so the intersection is the only target that
                  means "loadable whichever revision your toolchain
                  implements".

                  That establishes metamodel validity only: it is not a claim
                  of IDTA conformance, and it asserts nothing about whether a
                  submodel matches a published submodel template. Note also
                  that no AAS JSON Schema sets `additionalProperties`, so
                  schema validity alone cannot rule out a member the metamodel
                  does not define; `dpp-core` gates that separately.

                  **Unsigned, and it says so in a header.** This is a derived
                  representation of the signed canonical public view, which is
                  what `application/ld+json` returns for this same URL. The
                  public proof covers that payload, not this serialisation of
                  it, so attaching the signature here would hand a verifier a
                  proof that fails against the bytes it arrived with.

                  Every `200` therefore carries:

                  ```
                  Link: <{resolverBase}/dpp/{dppId}>; rel="alternate"; type="application/ld+json"
                  ```

                  `alternate` rather than `canonical`: the two representations
                  share one URL and are separated only by `Accept`, so a
                  `canonical` relation would point this resource at itself.
                  Follow the link with that `Accept` to obtain the signed
                  payload and its proof.

                  `resolverBase` is per-deployment (`RESOLVER_BASE_URL`,
                  default `https://id.odal-node.io`). Error responses carry no
                  `Link` — an error is not a representation of the passport.
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: |
        Machine-to-machine access via API keys. Keys are prefixed with
        `odal_sk_` and sent as `Authorization: Bearer odal_sk_...`.

        Keys are stored as SHA-256 hashes in PostgreSQL with a 12-character
        prefix index for fast lookup. The plain-text key is only returned
        once at creation time via `POST /vault/api/v1/api-keys`.
    BasicAuth:
      type: http
      scheme: basic
      description: |
        The operator's own bootstrap credential, from the `ADMIN_USERNAME` and
        `ADMIN_PASSWORD` environment variables. It mints the first API key on a
        fresh node — before any key exists — and is the lockout-recovery path
        afterwards, since it carries no `keyId` and so can revoke any key.

        Active in every environment where both variables are set; there is no
        production gate. Leave them unset once an API key exists if you do not
        want the path available.

        Reached only via the `Basic` scheme. A `Bearer` token is never matched
        against it, even one carrying the same `base64(user:pass)` payload.
    MutualTLS:
      type: mutualTLS
      description: |
        Client-certificate (mTLS) authentication for service-to-service calls.
        The standalone identity service accepts the internal `/internal/*`
        endpoints only from a client presenting `CN=odal-vault`. The fused
        `dpp-node` never exposes these over the network — it signs in-process.
  schemas:
    ManufacturerInfo:
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
          example: EcoTech GmbH
        address:
          type: string
          example: Hauptstraße 1, 10115 Berlin
        didWebUrl:
          type: string
          format: uri
          description: did:web URL for the manufacturer
          example: https://ecotech.example.com/.well-known/did.json
    MaterialEntry:
      type: object
      required:
        - name
        - weightKg
      properties:
        name:
          type: string
          example: Lithium carbonate
        weightKg:
          type: number
          description: Weight in kilograms
          example: 12.5
        recycledPct:
          type: number
          minimum: 0
          maximum: 100
          description: Percentage of recycled content (0.0–100.0)
          example: 35
        countryOfOrigin:
          type: string
          minLength: 2
          maxLength: 2
          description: ISO 3166-1 alpha-2 country code
          example: DE
...
    DppId:
      type: string
      format: uuid
      description: UUID v7 identifier assigned on creation. Embedded in QR codes and public URLs.
      example: 019723f4-1a2b-7c3d-8e4f-5a6b7c8d9e0f
    CarbonFootprint:
      type: object
      description: A CO₂-equivalent figure with the LCA context needed to read it. A bare number is not comparable across products — the lifecycle stage and system boundary are what make two figures mean the same thing.
      required:
        - valueKg
      properties:
        valueKg:
          type: number
          description: CO₂-equivalent value in kg per functional unit.
          example: 45.2
        lifecycleStage:
          type: string
          description: LCA lifecycle stage this figure covers.
          enum:
            - cradle-to-gate
            - cradle-to-grave
            - cradle-to-cradle
            - gate-to-grave
            - other
        systemBoundary:
          type: string
          description: LCA system-boundary standard used.
          enum:
            - EN-15804
            - ISO-14044
            - GHG-protocol
            - other
        methodologyRef:
          type: string
          description: Citation for the methodology behind the figure.
        performanceClass:
          type: string
          maxLength: 8
          description: Performance class label as defined by the applicable delegated act (e.g. an A–G band). Free text because the banding is per product group.
          example: B
...
    PassportResponse:
      type: object
      description: The canonical Digital Product Passport record. Fields marked required are emitted on every read; the rest are omitted when unset rather than sent as null, so a consumer must treat absence and null as the same thing.
      required:
        - id
        - productName
        - productGroup
        - manufacturer
        - materials
        - status
        - schemaVersion
        - createdAt
        - updatedAt
        - retentionLocked
        - version
      properties:
        id:
          $ref: '#/components/schemas/DppId'
        batchId:
          type:
            - string
            - 'null'
        productName:
          type: string
          example: EcoCell Pro 48V
        productGroup:
          type: string
          description: 'EU ESPR product group — the delegated-act bucket selecting the applicable schema and plugin. Deliberately an open string, not a closed enum: adding a product group is a catalog manifest plus a schema, not a release, and a product group this build does not know still round-trips its wire tag verbatim.'
          example: battery
        applicableInstruments:
          type: array
          description: |-
            The legal instruments this passport was issued under, fixed when the product was placed on the market and never recomputed afterwards.

            A set rather than a single value because acts accumulate: ESPR Art. 5(7) lets a group-specific delegated act supplement a horizontal one and the Regulation states no precedence rule, so the governing law is the union of the members' requirements.

            Not derivable from `productGroup`. A horizontal act can reach a product whose product group is not one the catalog models, so an entry may be asserted by the economic operator rather than resolved from the catalog — which `recorded` distinguishes. Read it as who asserted the entry, not as how much to trust it.
          items:
            type: object
            required:
              - instrument
              - recorded
            properties:
              instrument:
                type: string
                description: The instrument's catalog id.
                example: battery-reg-2023-1542
              recorded:
                type: string
                enum:
                  - catalog
                  - operator
                description: '`catalog` — resolved from the instrument catalog at issuance. `operator` — asserted by the economic operator placing the product on the market.'
                example: catalog
            additionalProperties: false
        granularity:
          type: string
          enum:
            - model
            - batch
            - item
          description: 'The level this passport describes. Set by the applicable delegated act (ESPR Art. 9(2)(d)), so it is a property of the law rather than a choice made here, and absent while no adopted act has fixed one — which is every product group today. Absent is not `item`: the EU registry registers batteries per item, but that is the registry''s operational position, not a level any act has set.'
          example: item
        manufacturer:
          $ref: '#/components/schemas/ManufacturerInfo'
        materials:
          type: array
          items:
            $ref: '#/components/schemas/MaterialEntry'
        co2ePerUnit:
          description: 'CO₂e per unit — manufacturer-supplied or engine-calculated. An object, not a bare number: a figure without its lifecycle stage and system boundary is not comparable to another product''s.'
          anyOf:
            - $ref: '#/components/schemas/CarbonFootprint'
            - type: 'null'
        repairabilityScore:
          description: Non-regulatory repairability heuristic — see the schema's own note.
          anyOf:
            - $ref: '#/components/schemas/RepairabilityScore'
            - type: 'null'
        complianceResult:
          allOf:
            - $ref: '#/components/schemas/ComplianceResult'
          description: The computed compliance determination, attached at create/update. Part of the signed payload and immutable after retention lock. Absent until a determination has been computed.
        lintResult:
          allOf:
            - $ref: '#/components/schemas/LintResult'
          description: Advisory plausibility findings. Absent until a lint pass has run, and recomputable after publish — unlike `complianceResult`.
        productGroupData:
          description: Explicitly `null` — not omitted — for a passport whose product group data has not been supplied yet, which is every draft created without it. The field has no `skip_serializing_if`, so it is always present on the wire.
          anyOf:
            - $ref: '#/components/schemas/ProductGroupData'
            - type: 'null'
        status:
          $ref: '#/components/schemas/PassportStatus'
        qrCodeUrl:
          type:
            - string
            - 'null'
          format: uri
          description: 'GS1 Digital Link the carrier (QR) encodes, set on publish: {resolverBase}/01/{gtin}/21/{serial} for a trade item, else {resolverBase}/dpp/{id}. resolverBase is per-deployment (RESOLVER_BASE_URL, default https://id.odal-node.io).'
        jwsSignature:
          type:
            - string
            - 'null'
          description: Compact JWS (Ed25519) over the **full** canonical payload. Null until published.
        publicJwsSignature:
          type: string
          description: Compact JWS over the **public (redacted) view**, so anyone can verify the public passport independently — the resolver checks this on the unauthenticated public route. Set at publish; absent for drafts.
        disclosureSignatures:
          type: object
          additionalProperties:
            type: string
          description: 'Compact JWS signatures over the **non-public** redacted views, keyed by disclosure set (e.g. `public+restricted+individual`), never by audience name. Every audience receiving more than the public view needs a proof over *its* view: `publicJwsSignature` covers only the public payload and `jwsSignature` only the full one, so a reader handed a filtered body and either of those holds a signature that cannot verify against the bytes it received. Empty for drafts.'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type:
            - string
            - 'null'
          format: date-time
        placedOnMarketDate:
          type: string
          format: date
          description: The date the product was placed on the EU market — the regulated triggering event fixing **which law governs it**, distinct from the three lifecycle dates above, none of which selects a rule. Staged EU obligations attach at placing on the market and do not move afterwards. Absence means the date was not declared; it is **not** licence to substitute the current date, and a determination depending on it has no answer.
        schemaVersion:
          type: string
          description: Semantic version of the **product group** schema `productGroupData` was validated against. Scoped to `productGroupData` only — the envelope fields have no equivalent version and never will, because they are shared by every product group's stored documents. The envelope's compatibility rule is additive only, permanently.
          example: 1.0.0
        retentionLocked:
          type: boolean
          description: |
            Set to `true` permanently on first publish. Retention-locked
            passports must remain accessible for the EU ESPR retention period.
        version:
          type: integer
          minimum: 1
          description: Monotonic version counter. `1` on first publish; incremented on the successor each time a new version supersedes this record.
        supersedesId:
          allOf:
            - $ref: '#/components/schemas/DppId'
          description: The passport this record supersedes. Absent for first versions.
        parentPassportRef:
          allOf:
            - $ref: '#/components/schemas/PassportRef'
          description: Cross-operator reference to the predecessor this passport derives from (second-life successor linkage).
        componentRefs:
          type: array
          description: 'Cross-operator references to the constituent passports this product is assembled from — its bill of materials. The inverse edge of `parentPassportRef`: these point down to many constituents, that points up to one predecessor.'
          items:
            $ref: '#/components/schemas/PassportRef'
        retentionUntil:
          type: string
          format: date-time
          description: Deadline by which this record must remain accessible, computed at publish from the product group's retention period. Regulation (EU) 2024/1781 Art. 9(2)(i) requires the delegated act to specify a period corresponding to at least the product's expected lifetime; Art. 11(e) restates it as an essential requirement, available including after the responsible operator's insolvency, liquidation or cessation of activity.
        productId:
          type: string
          format: uuid
          description: Opaque link to an internal product-template record. Not a legal identifier.
        commodityCode:
          type: string
          description: Customs tariff classification — HS-6, CN-8 or TARIC-10. Absent where the product group does not call for one; this node will not invent a classification it cannot derive.
          example: '85076000'
        operatorIdentifier:
          type: string
          description: 'EORI or national economic-operator identifier for the responsible party (Regulation (EU) 2024/1781 Annex III(k); issuance mechanics in Art. 12). **Frozen at publish — this is the operator that published the passport, not necessarily the one responsible for it now.** A transfer of responsibility does not rewrite it and cannot: published content is immutable and covered by the signature over it. For current responsibility, read the transfer chain.'
          example: DE123456789
        facility:
          allOf:
            - $ref: '#/components/schemas/FacilitySnapshot'
          description: Snapshot of the Annex III facility where this product was manufactured or processed, copied by value at create time.
        seal:
          allOf:
            - $ref: '#/components/schemas/SealedEnvelope'
          description: The eIDAS electronic seal applied to this passport. Absent until a seal has been applied; check its `placeholder` flag rather than inferring validity from presence.
...
  responses:
...
    NotFound:
      description: Resource not found within the operator's scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://problems.odal-node.io/not-found
            title: Not Found
            status: 404
            detail: DPP not found.
...
    NotAcceptable:
      description: |
        No representation matches the request's `Accept` header. The response
        body names the media types this resource can produce.

        A passport carrying no GTIN — an unsold-goods report, or an untyped
        product group — also gets this for `application/aas+json`: it identifies no
        trade item, so it has no AAS asset identity and therefore no AAS
        representation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://problems.odal-node.io/not-acceptable
            title: Not Acceptable
            status: 406
            detail: No representation matches 'application/pdf'. This resource is available as text/html, application/ld+json, or application/aas+json.